shuting/page.tsxのリファクタリング
